further size_t patches
git-svn-id: svn+ssh://svn.cprover.org/srv/svn/cbmc/trunk@3582 6afb6bc1-c8e4-404c-8f48-9ae832c5b171

+++ src/util/cout_message.cpp
@@ -56,7 +56,7 @@ void console_message_handlert::print(
 
     WriteConsoleW(
       out_handle, wide_message.c_str(),
-      wide_message.size(), &number_written, NULL);
+      (DWORD)wide_message.size(), &number_written, NULL);
 
     WriteConsoleW(out_handle, L"\r\n", 2, &number_written, NULL);
   }
